+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/enumeration/Role.kt
@@ -1,0 +1,5 @@
+package com.flabedu.blackpostoffice.commom.enumeration
+
+enum class Role {
+    NONE, USER, ADMIN
+}
