+++ src/main/kotlin/com/flabedu/blackpostoffice/controller/PostController.kt
@@ -1,8 +1,8 @@
 package com.flabedu.blackpostoffice.controller
 
 import com.flabedu.blackpostoffice.commom.annotation.AuthorizedAccessCheck
+import com.flabedu.blackpostoffice.commom.enumeration.Role.USER
 import com.flabedu.blackpostoffice.model.post.Post
-import com.flabedu.blackpostoffice.model.user.UserSignUp.Role.USER
 import com.flabedu.blackpostoffice.service.PostService
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.DeleteMapping
