+++ src/main/kotlin/com/flabedu/blackpostoffice/model/user/UserSignUp.kt
@@ -28,10 +28,7 @@ data class UserSignUp(
 
     val profileImagePath: String?,
 
-) {
-    enum class Role {
-        NONE, USER, ADMIN
-    }
+    ) {
 
     fun toPasswordEncryption(password: String) =
         UserSignUp(
